Remove cache
A key of the cache is "dateSpan" whose "until" field is set by time.Now
method which can handle nanoseconds. Since each goroutine is not
executed at exactly the same time, this causes cache misses even if the
date of dateSpan is matched. In this case, cache makes no sense and it
extends execution time wastefully. Moreover, even though fields of
dateSpan is changed to use dates, cache misses may frequently happen
depending on user's config. In order to shorten execution time, it seems
to be better not to use cache.